+++ lib/ui/page/home/page/chat/widget/chat_forward.dart
@@ -832,7 +832,7 @@ class _ChatForwardWidgetState extends State<ChatForwardWidget> {
                               ContextMenuButton(
                                 key: const Key('CopyButton'),
                                 label: PlatformUtils.isMobile
-                                    ? 'btn_copy_text_short'.l10n
+                                    ? 'btn_copy'.l10n
                                     : 'btn_copy_text'.l10n,
                                 trailing: SvgLoader.asset(
                                   'assets/icons/copy_small.svg',
@@ -843,8 +843,8 @@ class _ChatForwardWidgetState extends State<ChatForwardWidget> {
                             ContextMenuButton(
                               key: const Key('ReplyButton'),
                               label: PlatformUtils.isMobile
-                                  ? 'btn_reply_short'.l10n
-                                  : 'btn_reply'.l10n,
+                                  ? 'btn_reply'.l10n
+                                  : 'btn_reply_message'.l10n,
                               trailing: SvgLoader.asset(
                                 'assets/icons/reply.svg',
                                 height: 18,
@@ -854,8 +854,8 @@ class _ChatForwardWidgetState extends State<ChatForwardWidget> {
                             ContextMenuButton(
                               key: const Key('ForwardButton'),
                               label: PlatformUtils.isMobile
-                                  ? 'btn_forward_short'.l10n
-                                  : 'btn_forward'.l10n,
+                                  ? 'btn_forward'.l10n
+                                  : 'btn_forward_message'.l10n,
                               trailing: SvgLoader.asset(
                                 'assets/icons/forward.svg',
                                 height: 18,
@@ -899,7 +899,7 @@ class _ChatForwardWidgetState extends State<ChatForwardWidget> {
                               ),
                             ContextMenuButton(
                               label: PlatformUtils.isMobile
-                                  ? 'btn_delete_message_short'.l10n
+                                  ? 'btn_delete'.l10n
                                   : 'btn_delete_message'.l10n,
                               trailing: SvgLoader.asset(
                                 'assets/icons/delete_small.svg',