+++ lib/ui/page/home/widget/chat_tile.dart
@@ -80,7 +80,7 @@ class ChatTile extends StatelessWidget {
       key: Key('ChatTile_${chat?.chat.value.id}'),
       preventContextMenu: false,
       actions: actions,
-      darkenable: true,
+      indicateOpenedMenu: true,
       child: SizedBox(
         height: height,
         child: Padding(