+++ lib/ui/page/home/widget/contact_tile.dart
@@ -115,7 +115,7 @@ class ContactTile extends StatelessWidget {
           : null,
       preventContextMenu: preventContextMenu,
       actions: actions ?? [],
-      darkenable: true,
+      indicateOpenedMenu: true,
       child: Padding(
         padding: margin,
         child: InkWellWithHover(